requirements.txt
streamlit
pandas
supabase
fpdf2